refactor(accounting,payroll): classes dupliquées au même nom supprimées — DWS fusionné + PayrollRegularizationCalculator (Closes #6572)
