+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - **docs(fuel): mobile pompiste FuelStation — contrats backend + gel UI (BC-15 FUEL, Closes #5807).** FUEL-013 : `docs/mobile/FUEL_MOBILE.md` documente la surface pompiste déjà en production sur `main` (`/fuel-station/me/shifts|presence|cash-sessions|sales`, saisies vente/relevé/caisse, résolution par employee_id, isolation tenant). L'écran mobile (UI Flutter) reste **gelé** par le freeze 60 jours (FREEZE_SCOPE_60J.md, « apps mobiles non-employee ») — intégration prévue au gate J60 ou via exception fondateur, avec file idempotente hors-ligne (pattern RESTO-804/TRAVEL-704). Aucun code de production modifié — livrable documentation + preuve de couverture backend.
 
 ## [Unreleased]
+- **refactor(accounting,payroll): classes dupliquées au même nom supprimées (Closes #6572).** Accounting — `DocumentWorkflowService` fusionné : les méthodes `transition()` et `linkCreditNote()` de l'implémentation Infrastructure sont intégrées à `Accounting\Application\Services\DocumentWorkflowService` (source unique du workflow #5223, matrice TRANSITIONS + règles métier conservées), consommateurs migrés (`SeedAccountingDemoData` avec numérotation injectée, `DocumentWorkflowNumberingTest`, `AccountingDemoE2ETest`), doublon `Infrastructure\Services\DocumentWorkflowService.php` supprimé. Payroll — l'ancien `Infrastructure\Services\PayrollRegularizationService` (calcul du différentiel) est renommé `PayrollRegularizationCalculator` : plus de résolution de namespace implicite depuis `PayrollCalculator` (un `use` ajouté n'aurait plus basculé silencieusement sur la classe Application) ; `Application\Services\PayrollRegularizationService` reste la classe canonique (création/gestion des runs). Critère : 0 paire de classes même nom court dans deux namespaces vivants.
+
 - **test(restaurant): specs E2E Playwright du POS portail tenant (RESTO-902, #6231).** `front/web/e2e/restaurant-pos.spec.ts` (branche `bc/bc25-restaurant-e2e-pos`, mock API déterministe) : golden flow GJ-RESTO-01 complet (caisse → commande → ajout article → soumission → confirmation cuisine → addition → encaissement espèces → commande payée) + filtre catalogue `is_available=false`. Vérifié localement : Playwright chromium 2/2 verts, eslint 0.
 - **feat(restaurant): écran POS — prise de commande & encaissement (BC-25 RESTAURANT, Closes #6218).** RESTO-705 : page `front/web/src/app/(dashboard)/restaurant/pos/page.tsx` (portail tenant, spec §4.1) — ouverture de caisse (`pos-sessions`), nouvelle commande dine_in (`orders`), ajout d'articles depuis le catalogue groupé par catégories (`categories`/`products`), annulation d'article, soumission + confirmation (cuisine), addition (`bill`), encaissement espèces (`pay`, montant serveur), rafraîchissement de la commande après chaque action. Module `restaurant` dans `CLIENT_MODULES` (gating flag `restaurantmanager`, routes `/restaurant/pos` + `/restaurant/kitchen`). i18n fr/en/ar/tr (`restaurant.pos.*`, sync-web). Tests Jest 3/3 (catalogue + création de commande, flux complet article→soumission→confirmation→addition→encaissement, caisse fermée → ouverture proposée). Vérifié localement : tsc --noEmit, eslint --max-warnings 0, check-mojibake OK.
 - **feat(analytics): snapshots horodatés des read models de reporting (Closes #6243, BC-22-D10).** Infrastructure de snapshot prête à l'activation (déclencheur = dépassement du budget p95, jamais préventive) : table tenant `accounting_reporting_snapshots` (migration additive/idempotente, clé unique `(company_id, report, période)`, version incrémentée UNIQUEMENT quand le contenu change — deux recomputes identiques → même version, idempotence), service `AccountingReportingSnapshotService`, job `RecomputeAccountingReportingSnapshotJob` (tenant-scoped `EnsureTenantContext`, retry borné, log corrélé), commande `accounting:reporting-snapshot` (async ou `--sync`), bloc `data.snapshot` (source/version/refreshed_at) exposé par `GET /accounting/dashboard` (fraîcheur au consommateur). Tests `AccountingReportingSnapshotTest` (idempotence, version, isolation cross-tenant, exposition API). Docs `docs/architecture/ANALYTICS_SNAPSHOTS.md` (activation, recompute, rollback). OpenAPI + miroir/SDK régénérés.
